+++ solution/day04/src/CHANGELOG.md
@@ -5,3 +5,4 @@ Improvements to the tests:
 4. Cleaned up test setup.
 5. "it should throw an exception when adding existing comment" - the test checked the exception but not the state of the article; asserted that the comment was not added despite the error.
 6. Improved test names to focus just on article comments (and not on `Article` as a whole)
+7. Added more thorough tests for multiple comments.